# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /call-requests:
    post:
      summary: Create a call request (ServiceNow data source only).
      operationId: createCallRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCallRequestRequest'
      responses:
        "201":
          description: Call request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCallRequestResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /call-requests/search:
    post:
      summary: Search call requests (ServiceNow data source only).
      operationId: searchCallRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCallRequestsRequest'
      responses:
        "200":
          description: Call requests matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCallRequestsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /call-requests/search-all:
    post:
      summary: Search call requests across all cases, filtered by assignee/state.
      description: >
        Returns a paginated list of call requests (ServiceNow data source only). Call
        requests are inherently case-related; this endpoint provides independent search
        and filtering capabilities across cases, distinct from /call-requests/search
        which is scoped to a single case.
      operationId: searchAllCallRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAllCallRequestsRequest'
      responses:
        "200":
          description: Call requests matching the supplied filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCallRequestsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /call-requests/{id}:
    patch:
      summary: Update a call request (supported by the backing data source only).
      description: >
        The target state selects the behaviour: customer/agent state
        transitions, scheduling (with meetingDate, durationInMinutes,
        assignee), rejection or cancellation (with cancellationReason), or
        conclusion with post-call notes (notes, plan, attendees, actionItems,
        actualDurationMin).
      operationId: updateCallRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCallRequestRequest'
      responses:
        "200":
          description: Call request updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCallRequestResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Number of results per page. Defaults to 20, capped at 50.
        offset:
          type: integer
          description: Zero-based offset into the result set.
...
    CallRequestState:
      type: object
      description: State of a call request.
      properties:
        id:
          type: string
          enum:
            - pending_on_customer
            - pending_on_wso2
            - scheduled
            - customer_rejected
            - wso2_rejected
            - canceled
            - notes_pending
            - concluded
          description: State enum key.
        label:
          type: string
          description: Human-readable state label.
    CallRequestCaseRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        number:
          type: string
    CreateCallRequestRequest:
      type: object
      required: [caseId, reason, utcTimes, durationInMinutes]
      additionalProperties: false
      properties:
        caseId:
          type: string
          format: uuid
          description: UUID of the case this call request belongs to.
        reason:
          type: string
          description: Reason for the call request.
        utcTimes:
          type: array
          minItems: 1
          items:
            type: string
          description: Preferred UTC times for the call.
        durationInMinutes:
          type: integer
          minimum: 1
          description: Duration of the call in minutes.
    CreateCallRequestResponse:
      type: object
      properties:
        message:
          type: string
        callRequest:
          type: object
          properties:
            id:
              type: string
              format: uuid
            createdOn:
              type: string
            createdBy:
              type: string
            state:
              $ref: '#/components/schemas/CallRequestState'
...
    SearchCallRequestsRequest:
      type: object
      required: [caseId]
      additionalProperties: false
      properties:
        caseId:
          type: string
          format: uuid
          description: UUID of the case to search call requests for.
        filters:
          type: object
          additionalProperties: false
          properties:
            states:
              type: array
              items:
                type: string
                enum:
                  - pending_on_customer
                  - pending_on_wso2
                  - scheduled
                  - customer_rejected
                  - wso2_rejected
                  - canceled
                  - notes_pending
                  - concluded
              description: Filter by one or more states.
        pagination:
          $ref: '#/components/schemas/Pagination'
    CallRequestView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        case:
          $ref: '#/components/schemas/CallRequestCaseRef'
        reason:
          type: string
        preferredTimes:
          type: array
          items:
            type: string
        durationMin:
          type: integer
        scheduleTime:
          type: string
        meetingLink:
          type: string
        createdOn:
          type: string
        updatedOn:
          type: string
        state:
          $ref: '#/components/schemas/CallRequestState'
        cancellationReason:
          type: string
        assignee:
          type: string
          description: Identifier of the engineer the call is assigned to.
        notes:
          type: string
          description: Notes captured from the concluded call.
        plan:
          type: string
        attendees:
          type: string
        actionItems:
          type: string
        actualDurationMin:
          type: integer
          description: Actual call duration in minutes.
    SearchCallRequestsResponse:
      type: object
      properties:
        callRequests:
          type: array
          items:
            $ref: '#/components/schemas/CallRequestView'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    CallRequestSort:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          enum: [createdOn, updatedOn, scheduleTime]
          default: updatedOn
          description: Field to sort by. Defaults to updatedOn when omitted.
        order:
          type: string
          enum: [asc, desc]
          default: desc
          description: Sort direction. Defaults to desc when omitted.
    SearchAllCallRequestsFilters:
      type: object
      additionalProperties: false
      properties:
        assignedUserIds:
          type: array
          items:
            type: string
            format: uuid
          description: Filter by the parent case's assigned user(s).
        states:
          type: array
          items:
            type: string
            enum:
              - pending_on_customer
              - pending_on_wso2
              - scheduled
              - customer_rejected
              - wso2_rejected
              - canceled
              - notes_pending
              - concluded
          description: Filter by one or more call request states.
    SearchAllCallRequestsRequest:
      type: object
      additionalProperties: false
      properties:
        filters:
          $ref: '#/components/schemas/SearchAllCallRequestsFilters'
        sortBy:
          $ref: '#/components/schemas/CallRequestSort'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UpdateCallRequestRequest:
      type: object
      required: [state]
      additionalProperties: false
      properties:
        state:
          type: string
          enum:
            - pending_on_customer
            - pending_on_wso2
            - scheduled
            - customer_rejected
            - wso2_rejected
            - canceled
            - notes_pending
            - concluded
          description: New state for the call request.
        cancellationReason:
          type: string
          description: Reason for cancellation or rejection.
        utcTimes:
          type: array
          items:
            type: string
          description: Updated preferred UTC times.
        durationInMinutes:
          type: integer
          minimum: 1
          description: Updated duration in minutes.
        meetingDate:
          type: string
          format: date-time
          description: Scheduled meeting start time as an RFC3339 UTC timestamp (used when scheduling).
        assignee:
          type: string
          description: Identifier of the engineer the call is assigned to.
        notes:
          type: string
          description: Notes captured from the concluded call.
        plan:
          type: string
          description: Plan discussed during the call.
        attendees:
          type: string
          description: List of call attendees.
        actionItems:
          type: string
          description: Action items agreed during the call.
        actualDurationMin:
          type: integer
          minimum: 1
          description: Actual call duration in minutes.
    UpdateCallRequestResponse:
      type: object
      properties:
        message:
          type: string
        callRequest:
          type: object
          properties:
            id:
              type: string
              format: uuid
            updatedOn:
              type: string
            updatedBy:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...